Improve the 'isARestaurantPage' condition
Instead of just search for the first occurence of the word restaurant
in the breadcrumb, now I look on all tags to see if any of them is
restaurant. Some pages has another tag in the first occurrence but
it doesn't mean it shouldn't be treated as a restaurant.

Sample http://www.yelp.com/biz/dangerous-dans-diner-toronto